requirements.txt
streamlit
PyPDF2
langchain
langchain-community
langchain-text-splitters
faiss-cpu
openai
python-dotenv
sentence-transformers

